Improve Namespaces
