+++ app/Http/Requests/Release/Departure/ApproveDepartureRelease.php
@@ -1,15 +1,17 @@
 <?php
 
-namespace App\Http\Requests\DepartureRelease;
+namespace App\Http\Requests\Release\Departure;
 
 use App\Rules\Controller\ControllerPositionValid;
 use Illuminate\Foundation\Http\FormRequest;
 
-class RejectDepartureRelease extends FormRequest
+class ApproveDepartureRelease extends FormRequest
 {
     public function rules()
     {
         return [
+            'expires_in_seconds' => 'required|integer|min:1',
+            'released_at' => 'present|nullable|date_format:Y-m-d H:i:s',
             'controller_position_id' => [
                 'required',
                 'integer',
